requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
anthropic==0.94.0
openai==2.31.0
chromadb==1.5.7
python-dotenv==1.2.2
slowapi==0.1.9
